+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -17,37 +17,58 @@ public struct RuleList {
         }
         list = tmpList
     }
+
+    internal func configuredRulesWithDictionary(dictionary: [String: AnyObject]) -> [Rule] {
+        var rules = [Rule]()
+        for ruleType in list.values {
+            let identifier = ruleType.description.identifier
+            if let ruleConfig = dictionary[identifier] {
+                do {
+                    let configuredRule = try ruleType.init(config: ruleConfig)
+                    rules.append(configuredRule)
+                } catch {
+                    queuedPrintError("Invalid config for '\(identifier)'. Falling back to default.")
+                    rules.append(ruleType.init())
+                }
+            } else {
+                rules.append(ruleType.init())
+            }
+        }
+        return rules
+    }
 }
 
-public let masterRuleList = RuleList( rules: ClosingBraceRule.self,
-                                      ColonRule.self,
-                                      CommaRule.self,
-                                      ConditionalBindingCascadeRule.self,
-                                      ControlStatementRule.self,
-                                      CustomRules.self,
-                                      CyclomaticComplexityRule.self,
-                                      EmptyCountRule.self,
-                                      FileLengthRule.self,
-                                      ForceCastRule.self,
-                                      ForceTryRule.self,
-                                      ForceUnwrappingRule.self,
-                                      FunctionBodyLengthRule.self,
-                                      FunctionParameterCountRule.self,
-                                      LeadingWhitespaceRule.self,
-                                      LegacyConstantRule.self,
-                                      LegacyConstructorRule.self,
-                                      LineLengthRule.self,
-                                      MissingDocsRule.self,
-                                      NestingRule.self,
-                                      OpeningBraceRule.self,
-                                      OperatorFunctionWhitespaceRule.self,
-                                      ReturnArrowWhitespaceRule.self,
-                                      StatementPositionRule.self,
-                                      TodoRule.self,
-                                      TrailingNewlineRule.self,
-                                      TrailingSemicolonRule.self,
-                                      TrailingWhitespaceRule.self,
-                                      TypeBodyLengthRule.self,
-                                      TypeNameRule.self,
-                                      ValidDocsRule.self,
-                                      VariableNameRule.self)
+public let masterRuleList = RuleList(rules:
+    ClosingBraceRule.self,
+    ColonRule.self,
+    CommaRule.self,
+    ConditionalBindingCascadeRule.self,
+    ControlStatementRule.self,
+    CustomRules.self,
+    CyclomaticComplexityRule.self,
+    EmptyCountRule.self,
+    FileLengthRule.self,
+    ForceCastRule.self,
+    ForceTryRule.self,
+    ForceUnwrappingRule.self,
+    FunctionBodyLengthRule.self,
+    FunctionParameterCountRule.self,
+    LeadingWhitespaceRule.self,
+    LegacyConstantRule.self,
+    LegacyConstructorRule.self,
+    LineLengthRule.self,
+    MissingDocsRule.self,
+    NestingRule.self,
+    OpeningBraceRule.self,
+    OperatorFunctionWhitespaceRule.self,
+    ReturnArrowWhitespaceRule.self,
+    StatementPositionRule.self,
+    TodoRule.self,
+    TrailingNewlineRule.self,
+    TrailingSemicolonRule.self,
+    TrailingWhitespaceRule.self,
+    TypeBodyLengthRule.self,
+    TypeNameRule.self,
+    ValidDocsRule.self,
+    VariableNameRule.self
+)