+++ Source/swiftlint/Commands/AutoCorrectCommand.swift
@@ -17,17 +17,19 @@ struct AutoCorrectCommand: CommandType {
     let function = "Automatically correct warnings and errors"
 
     func run(options: AutoCorrectOptions) -> Result<(), CommandantError<()>> {
-        var configuration = Configuration(commandLinePath: options.configurationFile)
-        configuration.rootPath = options.path.absolutePathStandardized()
+        let configuration = Configuration(commandLinePath: options.configurationFile,
+            rootPath: options.path, quiet: options.quiet)
         return configuration.visitLintableFiles(options.path, action: "Correcting",
-            useScriptInputFiles: options.useScriptInputFiles) { linter in
+            quiet: options.quiet, useScriptInputFiles: options.useScriptInputFiles) { linter in
             let corrections = linter.correct()
-            if !corrections.isEmpty {
+            if !corrections.isEmpty && !options.quiet {
                 let correctionLogs = corrections.map({ $0.consoleDescription })
                 queuedPrint(correctionLogs.joinWithSeparator("\n"))
             }
         }.flatMap { files in
-            queuedPrintError("Done correcting \(files.count) files!")
+            if !options.quiet {
+                queuedPrintError("Done correcting \(files.count) files!")
+            }
             return .Success()
         }
     }
@@ -37,17 +39,17 @@ struct AutoCorrectOptions: OptionsType {
     let path: String
     let configurationFile: String
     let useScriptInputFiles: Bool
+    let quiet: Bool
 
-    // swiftlint:disable:next line_length
-    static func create(path: String) -> (configurationFile: String) -> (useScriptInputFiles: Bool) -> AutoCorrectOptions {
-        return { configurationFile in { useScriptInputFiles in
-            self.init(path: path, configurationFile: configurationFile,
-                      useScriptInputFiles: useScriptInputFiles)
-        }}
+    // swiftlint:disable line_length
+    static func create(path: String) -> (configurationFile: String) -> (useScriptInputFiles: Bool) -> (quiet: Bool) -> AutoCorrectOptions {
+        return { configurationFile in { useScriptInputFiles in { quiet in
+            self.init(path: path, configurationFile: configurationFile, useScriptInputFiles: useScriptInputFiles, quiet: quiet)
+        }}}
     }
 
-    // swiftlint:disable:next line_length
     static func evaluate(mode: CommandMode) -> Result<AutoCorrectOptions, CommandantError<CommandantError<()>>> {
+        // swiftlint:enable line_length
         return create
             <*> mode <| Option(key: "path",
                 defaultValue: "",
@@ -58,5 +60,8 @@ struct AutoCorrectOptions: OptionsType {
             <*> mode <| Option(key: "use-script-input-files",
                 defaultValue: false,
                 usage: "read SCRIPT_INPUT_FILE* environment variables as files")
+            <*> mode <| Option(key: "quiet",
+                defaultValue: false,
+                usage: "don't print status logs like 'Correcting <file>' & 'Done correcting'")
     }
 }